src/requirements.txt
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.1
elevenlabs==0.2.26
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.0.2
idna==3.3
ipython==8.15.0
itsdangerous==2.0.1
jedi==0.19.0
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
numpy==1.26.0
openai==0.19.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pycodestyle==2.8.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SpeechRecognition==3.10.0
stack-data==0.6.2
toml==0.10.2
tqdm==4.62.3
traitlets==5.10.0
typing_extensions==4.8.0
urllib3==1.26.7
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.0.2
